+++ CHANGELOG.md
@@ -11,6 +11,7 @@ at anytime.
 ### Changed
  * Remove unused upload_allowed option
  * Remove code related to packaging as that step is now done in the electron client
+ * Remove lbryum version check; use lbry-electron as version source
 
 ## [0.8.3] - 2017-02-15
 ### Fixed